requirements.txt
requests
python-taiga